+++ R/opq.R
@@ -155,6 +155,44 @@ add_osm_feature <- function (opq, key, value, key_exact = TRUE,
     opq
 }
 
+#' Add a feature specified by OSM ID to an Overpass query
+#'
+#' @param id One or more official OSM identifiers (long-form integers)
+#' @param type Type of object; must be either `node`, `way`, or `relation`
+#' @return \code{opq} object
+#' 
+#' @references
+#' \url{https://wiki.openstreetmap.org/wiki/Overpass_API/Overpass_QL#By_element_id}
+#'
+#' @export
+#'
+#' @examples
+#' \dontrun{
+#' id <- c (1489221200, 1489221321, 1489221491)
+#' dat <- opq_osm_id (type = "node", id = id) %>%
+#'     opq_string () %>%
+#'     osmdata_sf () %>%
+#'     .$osm_points
+#' id <- c (136190595, 136190596)
+#' dat <- opq_osm_id (type = "way", id = id) %>%
+#'     opq_string () %>%
+#'     osmdata_sf () %>%
+#'     .$osm_lines
+#' }
+opq_osm_id <- function (id = NULL, type = NULL)
+{
+    if (is.null (type))
+        stop ('type must be specified: one of node, way, or relation')
+    type <- match.arg (tolower (type), c ('node', 'way', 'relation'))
+
+    opq <- opq (1:4)
+    opq$bbox <- NULL
+    opq$features <- NULL
+    opq$id <- list (type = type, id = id)
+
+    opq
+}
+
 #' @rdname add_osm_feature
 #' @export
 add_feature <- function (opq, key, value, key_exact = TRUE,
@@ -181,10 +219,22 @@ add_feature <- function (opq, key, value, key_exact = TRUE,
 #' opq_string (q)
 opq_string <- function (opq)
 {
-  features <- paste (opq$features, collapse = '')
-  features <- paste0 (sprintf (' node %s (%s);\n', features, opq$bbox),
-                      sprintf (' way %s (%s);\n', features, opq$bbox),
-                      sprintf (' relation %s (%s);\n\n', features,
-                               opq$bbox))
-  paste0 (opq$prefix, features, opq$suffix)
+    res <- NULL
+    if (!is.null (opq$features))
+    {
+        features <- paste (opq$features, collapse = '')
+        features <- paste0 (sprintf (' node %s (%s);\n', features, opq$bbox),
+                            sprintf (' way %s (%s);\n', features, opq$bbox),
+                            sprintf (' relation %s (%s);\n\n', features,
+                                     opq$bbox))
+        res <- paste0 (opq$prefix, features, opq$suffix)
+    } else if (!is.null (opq$id))
+    {
+        id <- paste (opq$id$id, collapse = ',')
+        id <- sprintf(' %s(id:%s);\n', opq$id$type, id)
+        res <- paste0 (opq$prefix, id, opq$suffix)
+    }
+    if (is.null (res))
+        res <- paste0 (opq$prefix, opq$suffix) # to ensure a non-null return
+    return (res)
 }